+++ CHANGELOG.md
@@ -3,4 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-- [pkg](./packages/pkg/CHANGELOG.md)
+- [@vitery/copy](./packages/@vitery/copy/CHANGELOG.md)
+- [@vitery/html](./packages/@vitery/html/CHANGELOG.md)
+- [@vitery/svgr](./packages/@vitery/svgr/CHANGELOG.md)
+- [vitery](./packages/vitery/CHANGELOG.md)